Use pure applescript to open the target application
Using a shell script appears to be unnecessary and just adds complexity.

One possible issue with this approach is whether any applications depend
on environment information from the shell, but such issues would
probably arise when opening applications normally. The only application
I'm currently aware of with issues like this is Atom with the Go plugin,
which complains about missing environment variables and recommends using
the command line tool to launch it.