Using AppleScript to create internal OS X alias to original application
This structure makes the alias application continue to open the target
application even if the target application is moved to a different path.

In order to accomplish this, AppleScript is used to create an OS X alias
to the original application, and the alias is bundled within the
application. The launcher script then runs the open command on the
internal alias, which then opens the target application.